refactor: split gather proof paths into small and large modules
