--- jolt-atlas-core/src/onnx_proof/ops/gather.rs
@@ -1,991 +1,0 @@
-use crate::onnx_proof::{ops::OperatorProofTrait, ProofId, ProofType, Prover, Verifier};
-use atlas_onnx_tracer::{
-    model::{
-        trace::{LayerData, Trace},
-        ComputationGraph,
-    },
-    node::ComputationNode,
-    ops::{GatherLarge, GatherSmall, Operator},
-    tensor::Tensor,
-};
-use common::{CommittedPolynomial, VirtualPolynomial};
-use joltworks::{
-    config::{OneHotConfig, OneHotParams},
-    field::{IntoOpening, JoltField},
-    poly::{
-        eq_poly::EqPolynomial,
-        identity_poly::IdentityPolynomial,
-        multilinear_polynomial::{
-            BindingOrder, MultilinearPolynomial, PolynomialBinding, PolynomialEvaluation,
-        },
-        opening_proof::{
-            OpeningAccumulator, OpeningPoint, ProverOpeningAccumulator, SumcheckId,
-            VerifierOpeningAccumulator, BIG_ENDIAN, LITTLE_ENDIAN,
-        },
-        unipoly::UniPoly,
-    },
-    subprotocols::{
-        booleanity::{
-            BooleanitySumcheckParams, BooleanitySumcheckVerifier, SmallBooleanitySumcheckProver,
-        },
-        hamming_booleanity::{
-            HammingBooleanitySumcheckParams, HammingBooleanitySumcheckProver,
-            HammingBooleanitySumcheckVerifier,
-        },
-        hamming_weight::{
-            HammingWeightSumcheckParams, HammingWeightSumcheckProver, HammingWeightSumcheckVerifier,
-        },
-        shout::{self, RaOneHotEncoding},
-        sumcheck::{BatchedSumcheck, Sumcheck, SumcheckInstanceProof},
-        sumcheck_prover::SumcheckInstanceProver,
-        sumcheck_verifier::{SumcheckInstanceParams, SumcheckInstanceVerifier},
-    },
-    transcripts::Transcript,
-    utils::{errors::ProofVerifyError, math::Math, thread::unsafe_allocate_zero_vec},
-};
-use rayon::{
-    iter::{
-        IndexedParallelIterator, IntoParallelIterator, IntoParallelRefIterator,
-        IntoParallelRefMutIterator, ParallelIterator,
-    },
-    slice::ParallelSlice,
-};
-
-impl<F: JoltField, T: Transcript> OperatorProofTrait<F, T> for GatherSmall {
-    #[tracing::instrument(skip_all, name = "GatherSmall::prove")]
-    fn prove(
-        &self,
-        node: &ComputationNode,
-        prover: &mut Prover<F, T>,
-    ) -> Vec<(ProofId, SumcheckInstanceProof<F, T>)> {
-        let mut results = Vec::new();
-
-        let params = GatherParams::new(
-            node.clone(),
-            &prover.preprocessing.model.graph,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-        let mut exec_sumcheck = GatherProver::initialize(
-            &prover.trace,
-            params,
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-        let (proof, _) = Sumcheck::prove(
-            &mut exec_sumcheck,
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-        results.push((ProofId(node.idx, ProofType::Execution), proof));
-
-        let (hb_sumcheck, bool_sumcheck) = build_small_stage2_provers::<F>(node, prover);
-        let mut instances: Vec<Box<dyn SumcheckInstanceProver<_, _>>> =
-            vec![Box::new(hb_sumcheck), Box::new(bool_sumcheck)];
-        let (stage2_proof, _) = BatchedSumcheck::prove(
-            instances.iter_mut().map(|v| &mut **v as _).collect(),
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-        results.push((ProofId(node.idx, ProofType::RaOneHotChecks), stage2_proof));
-
-        let mut hw_sumcheck = build_small_stage3_prover::<F>(node, prover);
-        let (hw_proof, _) = Sumcheck::prove(
-            &mut hw_sumcheck,
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-        results.push((ProofId(node.idx, ProofType::RaHammingWeight), hw_proof));
-
-        results
-    }
-
-    #[tracing::instrument(skip_all, name = "GatherSmall::verify")]
-    fn verify(
-        &self,
-        node: &ComputationNode,
-        verifier: &mut Verifier<'_, F, T>,
-    ) -> Result<(), ProofVerifyError> {
-        let proof = verifier
-            .proofs
-            .get(&ProofId(node.idx, ProofType::Execution))
-            .ok_or(ProofVerifyError::MissingProof(node.idx))?;
-        let exec_sumcheck = GatherVerifier::new(
-            node.clone(),
-            &verifier.preprocessing.model.graph,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        );
-        Sumcheck::verify(
-            proof,
-            &exec_sumcheck,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        )?;
-
-        let ra_one_hot_proof = verifier
-            .proofs
-            .get(&ProofId(node.idx, ProofType::RaOneHotChecks))
-            .ok_or(ProofVerifyError::MissingProof(node.idx))?;
-        let (hb_sumcheck, bool_sumcheck) = build_small_stage2_verifiers::<F>(node, verifier);
-        let instances: Vec<&dyn SumcheckInstanceVerifier<F, T>> =
-            vec![&hb_sumcheck, &bool_sumcheck];
-        BatchedSumcheck::verify(
-            ra_one_hot_proof,
-            instances,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        )?;
-
-        let hw_proof = verifier
-            .proofs
-            .get(&ProofId(node.idx, ProofType::RaHammingWeight))
-            .ok_or(ProofVerifyError::MissingProof(node.idx))?;
-        let hw_sumcheck = build_small_stage3_verifier::<F>(node, verifier);
-        Sumcheck::verify(
-            hw_proof,
-            &hw_sumcheck,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        )?;
-
-        Ok(())
-    }
-
-    fn get_committed_polynomials(&self, node: &ComputationNode) -> Vec<CommittedPolynomial> {
-        vec![CommittedPolynomial::GatherRa(node.idx)]
-    }
-}
-
-impl<F: JoltField, T: Transcript> OperatorProofTrait<F, T> for GatherLarge {
-    #[tracing::instrument(skip_all, name = "GatherLarge::prove")]
-    fn prove(
-        &self,
-        node: &ComputationNode,
-        prover: &mut Prover<F, T>,
-    ) -> Vec<(ProofId, SumcheckInstanceProof<F, T>)> {
-        let mut results = Vec::new();
-
-        // Execution proof (Read-Value and Raf)
-        let params = GatherParams::new(
-            node.clone(),
-            &prover.preprocessing.model.graph,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-        let mut exec_sumcheck = GatherProver::initialize(
-            &prover.trace,
-            params,
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-        let (proof, _) = Sumcheck::prove(
-            &mut exec_sumcheck,
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-        results.push((ProofId(node.idx, ProofType::Execution), proof));
-
-        // RA one-hot checks
-        let encoding = GatherRaEncoding::new(node);
-        let lookup_indices = gather_lookup_indices(node, &prover.trace);
-        let [ra_prover, hw_prover, bool_prover] = shout::ra_onehot_provers(
-            &encoding,
-            &lookup_indices,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-        let mut instances: Vec<Box<dyn SumcheckInstanceProver<_, _>>> =
-            vec![ra_prover, hw_prover, bool_prover];
-
-        let (stage2_proof, _) = BatchedSumcheck::prove(
-            instances.iter_mut().map(|v| &mut **v as _).collect(),
-            &mut prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        results.push((ProofId(node.idx, ProofType::RaOneHotChecks), stage2_proof));
-
-        results
-    }
-
-    #[tracing::instrument(skip_all, name = "GatherLarge::verify")]
-    fn verify(
-        &self,
-        node: &ComputationNode,
-        verifier: &mut Verifier<'_, F, T>,
-    ) -> Result<(), ProofVerifyError> {
-        // Execution verification (Read-Value and Raf)
-        let proof = verifier
-            .proofs
-            .get(&ProofId(node.idx, ProofType::Execution))
-            .ok_or(ProofVerifyError::MissingProof(node.idx))?;
-        let exec_sumcheck = GatherVerifier::new(
-            node.clone(),
-            &verifier.preprocessing.model.graph,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        );
-        Sumcheck::verify(
-            proof,
-            &exec_sumcheck,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        )?;
-
-        // RA one-hot verification
-        let ra_one_hot_proof = verifier
-            .proofs
-            .get(&ProofId(node.idx, ProofType::RaOneHotChecks))
-            .ok_or(ProofVerifyError::MissingProof(node.idx))?;
-        let encoding = GatherRaEncoding::new(node);
-        let [ra_verifier, hw_verifier, bool_verifier] =
-            shout::ra_onehot_verifiers(&encoding, &verifier.accumulator, &mut verifier.transcript);
-        let instances: Vec<&dyn SumcheckInstanceVerifier<F, T>> =
-            vec![&*ra_verifier, &*hw_verifier, &*bool_verifier];
-        BatchedSumcheck::verify(
-            ra_one_hot_proof,
-            instances,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        )?;
-
-        Ok(())
-    }
-
-    fn get_committed_polynomials(&self, node: &ComputationNode) -> Vec<CommittedPolynomial> {
-        let encoding = GatherRaEncoding::new(node);
-        let d = encoding.one_hot_params().instruction_d;
-        (0..d)
-            .map(|i| CommittedPolynomial::GatherRaD(node.idx, i))
-            .collect()
-    }
-}
-
-const DEGREE_BOUND: usize = 2;
-
-/// Parameters for proving gather (indexed lookup) operations.
-///
-/// Gather selects elements from an input dictionary tensor using an index tensor.
-/// Uses one-hot encoding to prove correct lookups with a folding challenge gamma.
-#[derive(Clone)]
-pub struct GatherParams<F: JoltField> {
-    gamma: F,
-    r_node_output: OpeningPoint<BIG_ENDIAN, F>,
-    computation_node: ComputationNode,
-    lookup_vars: usize,
-    num_words: usize,
-}
-
-impl<F: JoltField> GatherParams<F> {
-    /// Create new gather parameters from a computation node, graph, accumulator, and transcript.
-    pub fn new(
-        computation_node: ComputationNode,
-        graph: &ComputationGraph,
-        accumulator: &dyn OpeningAccumulator<F>,
-        transcript: &mut impl Transcript,
-    ) -> Self {
-        let gamma = transcript.challenge_scalar();
-
-        let dict_len = match &computation_node.operator {
-            Operator::GatherSmall(gather_op) => gather_op.dict_len,
-            Operator::GatherLarge(gather_op) => gather_op.dict_len,
-            _ => panic!("Expected Gather operator"),
-        };
-
-        let input_indices = &graph.nodes.get(&computation_node.inputs[1]).unwrap();
-        let num_words = dict_len.next_power_of_two();
-        let lookup_vars = input_indices.pow2_padded_num_output_elements().log_2();
-
-        let r_node_output = accumulator
-            .get_node_output_opening(computation_node.idx)
-            .0
-            .r;
-        Self {
-            gamma,
-            r_node_output: r_node_output.into(),
-            computation_node,
-            num_words,
-            lookup_vars,
-        }
-    }
-}
-
-impl<F: JoltField> SumcheckInstanceParams<F> for GatherParams<F> {
-    fn degree(&self) -> usize {
-        DEGREE_BOUND
-    }
-
-    fn input_claim(&self, accumulator: &dyn OpeningAccumulator<F>) -> F {
-        let rv_claim = accumulator
-            .get_node_output_opening(self.computation_node.idx)
-            .1;
-
-        let index_claim = accumulator
-            .get_virtual_polynomial_opening(
-                VirtualPolynomial::NodeOutput(self.computation_node.inputs[1]),
-                SumcheckId::NodeExecution(self.computation_node.idx),
-            )
-            .1;
-
-        rv_claim + self.gamma * index_claim
-    }
-
-    fn normalize_opening_point(&self, challenges: &[F]) -> OpeningPoint<BIG_ENDIAN, F> {
-        OpeningPoint::<LITTLE_ENDIAN, F>::new(challenges.to_vec()).match_endianness()
-    }
-
-    fn num_rounds(&self) -> usize {
-        self.num_words.log_2()
-    }
-}
-
-/// Prover state for gather sumcheck protocol.
-///
-/// Implements a Read-Raf sumcheck asserting that each output[i] corresponds to a lookup
-/// into the input dictionary at index input1[i].
-pub struct GatherProver<F: JoltField> {
-    params: GatherParams<F>,
-    dictionary: MultilinearPolynomial<F>,
-    index_onehot: MultilinearPolynomial<F>,
-    identity: IdentityPolynomial<F>,
-}
-
-struct GatherRaEncoding {
-    node_idx: usize,
-    index_input_idx: usize,
-    num_words: usize,
-}
-
-impl GatherRaEncoding {
-    fn new(computation_node: &ComputationNode) -> Self {
-        let gather_op = match &computation_node.operator {
-            Operator::GatherLarge(gather_op) => gather_op,
-            _ => panic!("Expected GatherLarge operator"),
-        };
-        Self {
-            node_idx: computation_node.idx,
-            index_input_idx: computation_node.inputs[1],
-            num_words: gather_op.dict_len.next_power_of_two(),
-        }
-    }
-}
-
-impl RaOneHotEncoding for GatherRaEncoding {
-    fn committed_poly(&self, d: usize) -> CommittedPolynomial {
-        CommittedPolynomial::GatherRaD(self.node_idx, d)
-    }
-
-    fn r_cycle_source(&self) -> (VirtualPolynomial, SumcheckId) {
-        (
-            VirtualPolynomial::NodeOutput(self.index_input_idx),
-            SumcheckId::NodeExecution(self.node_idx),
-        )
-    }
-
-    fn ra_source(&self) -> (VirtualPolynomial, SumcheckId) {
-        (
-            VirtualPolynomial::NodeOutputRa(self.node_idx),
-            SumcheckId::NodeExecution(self.node_idx),
-        )
-    }
-
-    fn log_k(&self) -> usize {
-        self.num_words.log_2()
-    }
-
-    fn one_hot_params(&self) -> OneHotParams {
-        OneHotParams::from_config_and_log_K(&OneHotConfig::default(), self.log_k())
-    }
-}
-
-fn gather_lookup_indices(computation_node: &ComputationNode, trace: &Trace) -> Vec<usize> {
-    let LayerData { operands, .. } = Trace::layer_data(trace, computation_node);
-    let [_, indexes] = operands[..] else {
-        panic!("Expected two operands for Gather operation")
-    };
-    indexes
-        .padded_next_power_of_two()
-        .data()
-        .par_iter()
-        .map(|&x| x as usize)
-        .collect()
-}
-
-fn build_small_stage2_provers<F: JoltField>(
-    computation_node: &ComputationNode,
-    prover: &mut Prover<F, impl Transcript>,
-) -> (
-    HammingBooleanitySumcheckProver<F>,
-    SmallBooleanitySumcheckProver<F>,
-) {
-    let Operator::GatherSmall(gather_op) = &computation_node.operator else {
-        panic!("Expected GatherSmall operator")
-    };
-    let LayerData { operands, .. } = Trace::layer_data(&prover.trace, computation_node);
-    let [dict, indexes] = operands[..] else {
-        panic!("Expected two operands for Gather operation")
-    };
-    let num_words = dict.dims()[gather_op.axis];
-    let num_lookups = indexes.len();
-
-    let hb_params = small_ra_hamming_bool_params::<F>(
-        computation_node,
-        num_lookups,
-        &prover.accumulator,
-        &mut prover.transcript,
-    );
-    let bool_params = small_ra_booleanity_params::<F>(
-        computation_node,
-        num_words,
-        num_lookups,
-        &prover.accumulator,
-        &mut prover.transcript,
-    );
-
-    let hw = {
-        let mut lookup_vec = vec![F::one(); indexes.len()];
-        lookup_vec.resize(indexes.len().next_power_of_two(), F::zero());
-        lookup_vec
-    };
-    let ra_evals = compute_ra_evals(&bool_params.r_cycle, indexes, num_words);
-    let indexes_u = indexes.iter().map(|&x| Some(x as u16)).collect();
-
-    let hb_sumcheck = HammingBooleanitySumcheckProver::gen(hb_params, vec![hw]);
-    let bool_sumcheck =
-        SmallBooleanitySumcheckProver::gen(bool_params, vec![ra_evals], vec![indexes_u]);
-
-    (hb_sumcheck, bool_sumcheck)
-}
-
-fn build_small_stage2_verifiers<F: JoltField>(
-    computation_node: &ComputationNode,
-    verifier: &mut Verifier<'_, F, impl Transcript>,
-) -> (
-    HammingBooleanitySumcheckVerifier<F>,
-    BooleanitySumcheckVerifier<F>,
-) {
-    let Operator::GatherSmall(gather_op) = &computation_node.operator else {
-        panic!("Expected GatherSmall operator")
-    };
-    let graph = &verifier.preprocessing.model.graph;
-    let dict = graph.nodes.get(&computation_node.inputs[0]).unwrap();
-    let indices = graph.nodes.get(&computation_node.inputs[1]).unwrap();
-
-    let num_words = dict.output_dims[gather_op.axis];
-    let num_lookups = indices.pow2_padded_num_output_elements();
-
-    let hb_params = small_ra_hamming_bool_params::<F>(
-        computation_node,
-        num_lookups,
-        &verifier.accumulator,
-        &mut verifier.transcript,
-    );
-    let bool_params = small_ra_booleanity_params::<F>(
-        computation_node,
-        num_words,
-        num_lookups,
-        &verifier.accumulator,
-        &mut verifier.transcript,
-    );
-
-    let hb_sumcheck = HammingBooleanitySumcheckVerifier::new(hb_params);
-    let bool_sumcheck = BooleanitySumcheckVerifier::new(bool_params);
-
-    (hb_sumcheck, bool_sumcheck)
-}
-
-fn small_ra_hamming_bool_params<F: JoltField>(
-    computation_node: &ComputationNode,
-    num_lookups: usize,
-    opening_accumulator: &dyn OpeningAccumulator<F>,
-    _transcript: &mut impl Transcript,
-) -> HammingBooleanitySumcheckParams<F> {
-    let polynomial_types = vec![VirtualPolynomial::HammingWeight];
-
-    let r_lookup = opening_accumulator
-        .get_virtual_polynomial_opening(
-            VirtualPolynomial::NodeOutput(computation_node.inputs[1]),
-            SumcheckId::NodeExecution(computation_node.idx),
-        )
-        .0
-        .r;
-
-    HammingBooleanitySumcheckParams {
-        d: 1,
-        num_rounds: num_lookups.log_2(),
-        gamma_powers: vec![F::one()],
-        polynomial_types,
-        r_cycle: r_lookup,
-        sumcheck_id: SumcheckId::RamHammingBooleanity,
-    }
-}
-
-fn small_ra_booleanity_params<F: JoltField>(
-    computation_node: &ComputationNode,
-    num_words: usize,
-    num_lookups: usize,
-    opening_accumulator: &dyn OpeningAccumulator<F>,
-    transcript: &mut impl Transcript,
-) -> BooleanitySumcheckParams<F> {
-    let polynomial_type = CommittedPolynomial::GatherRa(computation_node.idx);
-
-    let r_lookup = opening_accumulator
-        .get_virtual_polynomial_opening(
-            VirtualPolynomial::NodeOutput(computation_node.inputs[1]),
-            SumcheckId::NodeExecution(computation_node.idx),
-        )
-        .0
-        .r;
-    let r_address = transcript
-        .challenge_vector_optimized::<F>(num_words.log_2())
-        .into_opening();
-
-    BooleanitySumcheckParams {
-        d: 1,
-        log_k_chunk: num_words.log_2(),
-        log_t: num_lookups.log_2(),
-        r_cycle: r_lookup,
-        r_address,
-        polynomial_types: vec![polynomial_type],
-        gammas: vec![F::Challenge::from(1)],
-        sumcheck_id: SumcheckId::Booleanity,
-    }
-}
-
-fn build_small_stage3_prover<F: JoltField>(
-    computation_node: &ComputationNode,
-    prover: &mut Prover<F, impl Transcript>,
-) -> HammingWeightSumcheckProver<F> {
-    let Operator::GatherSmall(gather_op) = &computation_node.operator else {
-        panic!("Expected GatherSmall operator")
-    };
-    let LayerData { operands, .. } = Trace::layer_data(&prover.trace, computation_node);
-    let [dict, indexes] = operands[..] else {
-        panic!("Expected two operands for Gather operation")
-    };
-    let num_words = dict.dims()[gather_op.axis];
-
-    let hw_params = small_ra_hamming_weight_params::<F>(
-        computation_node,
-        num_words,
-        &prover.accumulator,
-        &mut prover.transcript,
-    );
-
-    let ra_evals = compute_ra_evals(&hw_params.r_cycle, indexes, num_words);
-
-    HammingWeightSumcheckProver::gen(hw_params, vec![ra_evals])
-}
-
-fn build_small_stage3_verifier<F: JoltField>(
-    computation_node: &ComputationNode,
-    verifier: &mut Verifier<'_, F, impl Transcript>,
-) -> HammingWeightSumcheckVerifier<F> {
-    let Operator::GatherSmall(gather_op) = &computation_node.operator else {
-        panic!("Expected GatherSmall operator")
-    };
-    let graph = &verifier.preprocessing.model.graph;
-    let dict = graph.nodes.get(&computation_node.inputs[0]).unwrap();
-
-    let num_words = dict.output_dims[gather_op.axis];
-
-    let hw_params = small_ra_hamming_weight_params::<F>(
-        computation_node,
-        num_words,
-        &verifier.accumulator,
-        &mut verifier.transcript,
-    );
-
-    HammingWeightSumcheckVerifier::new(hw_params)
-}
-
-fn small_ra_hamming_weight_params<F: JoltField>(
-    computation_node: &ComputationNode,
-    num_words: usize,
-    opening_accumulator: &dyn OpeningAccumulator<F>,
-    _transcript: &mut impl Transcript,
-) -> HammingWeightSumcheckParams<F> {
-    let polynomial_types = vec![CommittedPolynomial::GatherRa(computation_node.idx)];
-
-    let r_lookup = opening_accumulator
-        .get_virtual_polynomial_opening(
-            VirtualPolynomial::NodeOutput(computation_node.inputs[1]),
-            SumcheckId::NodeExecution(computation_node.idx),
-        )
-        .0
-        .r;
-
-    HammingWeightSumcheckParams {
-        d: 1,
-        num_rounds: num_words.log_2(),
-        gamma_powers: vec![F::one()],
-        polynomial_types,
-        sumcheck_id: SumcheckId::HammingWeight,
-        r_cycle: r_lookup,
-    }
-}
-
-impl<F: JoltField> GatherProver<F> {
-    /// Initialize the prover with trace data, parameters, accumulator, and transcript.
-    pub fn initialize(
-        trace: &Trace,
-        params: GatherParams<F>,
-        accumulator: &mut ProverOpeningAccumulator<F>,
-        transcript: &mut impl Transcript,
-    ) -> Self {
-        let LayerData { operands, .. } = Trace::layer_data(trace, &params.computation_node);
-        let [dictionary, indexes] = operands[..] else {
-            panic!("Expected two operands for Gather operation")
-        };
-
-        let dict_len = dictionary.dims()[0];
-        let word_dim = dictionary.dims().iter().product::<usize>() / dict_len;
-
-        let (r_index, r_word) = params.r_node_output.r.split_at(params.lookup_vars);
-        assert_eq!(r_word.len(), word_dim.log_2(),);
-
-        let padded_indexes = indexes.padded_next_power_of_two();
-        let index_claim = MultilinearPolynomial::from(padded_indexes.clone()).evaluate(r_index);
-        accumulator.append_virtual(
-            transcript,
-            VirtualPolynomial::NodeOutput(params.computation_node.inputs[1]),
-            SumcheckId::NodeExecution(params.computation_node.idx),
-            r_index.to_vec().into(),
-            index_claim,
-        );
-
-        let index_onehot: Vec<F> = compute_ra_evals(r_index, &padded_indexes, params.num_words);
-        let dict: Vec<F> = fold_dictionary(r_word, dictionary, params.num_words);
-
-        let index_onehot = MultilinearPolynomial::from(index_onehot);
-        let dictionary = MultilinearPolynomial::from(dict);
-        assert_eq!(index_onehot.len(), dictionary.len());
-        let identity = IdentityPolynomial::new(params.num_words.log_2());
-
-        #[cfg(test)]
-        {
-            let rv_claim = accumulator
-                .get_node_output_opening(params.computation_node.idx)
-                .1;
-            let claim = (0..index_onehot.len())
-                .map(|i| {
-                    let a = index_onehot.get_bound_coeff(i);
-                    let b = dictionary.get_bound_coeff(i);
-                    let int = F::from_u32(i as u32);
-                    a * (b + params.gamma * int)
-                })
-                .sum();
-            assert_eq!(rv_claim + params.gamma * index_claim, claim)
-        }
-        Self {
-            params,
-            index_onehot,
-            dictionary,
-            identity,
-        }
-    }
-}
-
-impl<F: JoltField, T: Transcript> SumcheckInstanceProver<F, T> for GatherProver<F> {
-    fn get_params(&self) -> &dyn SumcheckInstanceParams<F> {
-        &self.params
-    }
-
-    fn compute_message(&mut self, _round: usize, previous_claim: F) -> UniPoly<F> {
-        let Self {
-            index_onehot,
-            dictionary,
-            identity,
-            ..
-        } = self;
-
-        let univariate_poly_evals: [F; 2] = (0..index_onehot.len() / 2)
-            .into_par_iter()
-            .map(|i| {
-                let ra_evals =
-                    index_onehot.sumcheck_evals(i, DEGREE_BOUND, BindingOrder::LowToHigh);
-                let dict_evals =
-                    dictionary.sumcheck_evals(i, DEGREE_BOUND, BindingOrder::LowToHigh);
-                let id_evals = identity.sumcheck_evals(i, DEGREE_BOUND, BindingOrder::LowToHigh);
-
-                [
-                    ra_evals[0] * (dict_evals[0] + id_evals[0] * self.params.gamma),
-                    ra_evals[1] * (dict_evals[1] + id_evals[1] * self.params.gamma),
-                ]
-            })
-            .reduce(
-                || [F::zero(); 2],
-                |running, new| [running[0] + new[0], running[1] + new[1]],
-            );
-
-        UniPoly::from_evals_and_hint(previous_claim, &univariate_poly_evals)
-    }
-
-    fn ingest_challenge(&mut self, r_j: F::Challenge, _round: usize) {
-        self.index_onehot
-            .bind_parallel(r_j, BindingOrder::LowToHigh);
-        self.dictionary.bind_parallel(r_j, BindingOrder::LowToHigh);
-        self.identity.bind_parallel(r_j, BindingOrder::LowToHigh);
-    }
-
-    fn cache_openings(
-        &self,
-        accumulator: &mut ProverOpeningAccumulator<F>,
-        transcript: &mut T,
-        sumcheck_challenges: &[F::Challenge],
-    ) {
-        let opening_point = self
-            .params
-            .normalize_opening_point(&sumcheck_challenges.into_opening());
-
-        let (r_index, r_word) = self
-            .params
-            .r_node_output
-            .r
-            .split_at(self.params.lookup_vars);
-
-        let r_idx_onehot = [&opening_point.r, r_index].concat();
-        accumulator.append_virtual(
-            transcript,
-            VirtualPolynomial::NodeOutputRa(self.params.computation_node.idx),
-            SumcheckId::NodeExecution(self.params.computation_node.idx),
-            r_idx_onehot.into(),
-            self.index_onehot.final_sumcheck_claim(),
-        );
-        let r_dict = [&opening_point.r, r_word].concat();
-        accumulator.append_virtual(
-            transcript,
-            VirtualPolynomial::NodeOutput(self.params.computation_node.inputs[0]),
-            SumcheckId::NodeExecution(self.params.computation_node.idx),
-            r_dict.into(),
-            self.dictionary.final_sumcheck_claim(),
-        );
-    }
-}
-
-/// Verifier for gather sumcheck protocol.
-///
-/// Verifies that the prover's claims are consistent with the gather (indexed lookup) operation.
-pub struct GatherVerifier<F: JoltField> {
-    params: GatherParams<F>,
-}
-
-impl<F: JoltField> GatherVerifier<F> {
-    /// Create a new verifier for the gather operation, caching index operand openings.
-    pub fn new(
-        computation_node: ComputationNode,
-        graph: &ComputationGraph,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut impl Transcript,
-    ) -> Self {
-        let params = GatherParams::new(computation_node, graph, accumulator, transcript);
-
-        let (r_index, _) = params.r_node_output.r.split_at(params.lookup_vars);
-        accumulator.append_virtual(
-            transcript,
-            VirtualPolynomial::NodeOutput(params.computation_node.inputs[1]),
-            SumcheckId::NodeExecution(params.computation_node.idx),
-            r_index.to_vec().into(),
-        );
-
-        Self { params }
-    }
-}
-
-impl<F: JoltField, T: Transcript> SumcheckInstanceVerifier<F, T> for GatherVerifier<F> {
-    fn get_params(&self) -> &dyn SumcheckInstanceParams<F> {
-        &self.params
-    }
-
-    fn expected_output_claim(
-        &self,
-        accumulator: &VerifierOpeningAccumulator<F>,
-        sumcheck_challenges: &[F::Challenge],
-    ) -> F {
-        let opening_point = self
-            .params
-            .normalize_opening_point(&sumcheck_challenges.into_opening());
-
-        let ra_claim = accumulator
-            .get_virtual_polynomial_opening(
-                VirtualPolynomial::NodeOutputRa(self.params.computation_node.idx),
-                SumcheckId::NodeExecution(self.params.computation_node.idx),
-            )
-            .1;
-        let int_eval =
-            IdentityPolynomial::new(self.params.num_words.log_2()).evaluate(&opening_point.r);
-        let dict_claim = accumulator.get_node_output_claim(
-            self.params.computation_node.inputs[0],
-            self.params.computation_node.idx,
-        );
-        ra_claim * (dict_claim + self.params.gamma * int_eval)
-    }
-
-    fn cache_openings(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-        sumcheck_challenges: &[F::Challenge],
-    ) {
-        let opening_point = self
-            .params
-            .normalize_opening_point(&sumcheck_challenges.into_opening());
-
-        let (r_index, r_word) = self
-            .params
-            .r_node_output
-            .r
-            .split_at(self.params.lookup_vars);
-        let r_idx_onehot = [&opening_point.r, r_index].concat();
-        accumulator.append_virtual(
-            transcript,
-            VirtualPolynomial::NodeOutputRa(self.params.computation_node.idx),
-            SumcheckId::NodeExecution(self.params.computation_node.idx),
-            r_idx_onehot.into(),
-        );
-        let r_dict = [&opening_point.r, r_word].concat();
-        accumulator.append_virtual(
-            transcript,
-            VirtualPolynomial::NodeOutput(self.params.computation_node.inputs[0]),
-            SumcheckId::NodeExecution(self.params.computation_node.idx),
-            r_dict.into(),
-        );
-    }
-}
-
-// From the read indexes, computes the bound ra vector.
-fn compute_ra_evals<F>(r: &[F], indexes: &Tensor<i32>, num_words: usize) -> Vec<F>
-where
-    F: JoltField,
-{
-    let e = EqPolynomial::evals(r);
-    let num_threads = rayon::current_num_threads();
-    let chunk_size = indexes.len().div_ceil(num_threads);
-
-    let indexes_usize = indexes
-        .par_iter()
-        .map(|&x| x as usize)
-        .collect::<Vec<usize>>();
-
-    let partial_results: Vec<Vec<F>> = indexes_usize
-        .par_chunks(chunk_size)
-        .enumerate()
-        .map(|(chunk_idx, chunk)| {
-            let mut local_ra = unsafe_allocate_zero_vec::<F>(num_words);
-            let base_idx = chunk_idx * chunk_size;
-            chunk.iter().enumerate().for_each(|(local_j, &k)| {
-                let global_j = base_idx + local_j;
-                local_ra[k] += e[global_j];
-            });
-            local_ra
-        })
-        .collect();
-    let mut ra = unsafe_allocate_zero_vec::<F>(num_words);
-    for partial in partial_results {
-        ra.par_iter_mut()
-            .zip(partial.par_iter())
-            .for_each(|(dest, &src)| *dest += src);
-    }
-    ra
-}
-
-// TODO: Assert correct behavior for axis != 0
-fn fold_dictionary<F: JoltField>(r: &[F], dictionary: &Tensor<i32>, num_words: usize) -> Vec<F> {
-    let E = EqPolynomial::evals(r);
-
-    let mut folded: Vec<F> = dictionary
-        .par_chunks(E.len())
-        .map(|word_vector| {
-            word_vector
-                .iter()
-                .zip(E.iter())
-                .map(|(&word_coeff, &e)| F::from_i32(word_coeff) * e)
-                .sum()
-        })
-        .collect();
-    folded.resize(num_words, F::zero());
-    folded
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::onnx_proof::ops::test::unit_test_op;
-    use atlas_onnx_tracer::{
-        model::test::ModelBuilder, model::Model, ops::Operator, tensor::Tensor,
-    };
-    use rand::{rngs::StdRng, SeedableRng};
-
-    fn gather_model(input_shape: &[usize], dictionnary_len: usize, word_dim: usize) -> Model {
-        let mut b = ModelBuilder::new();
-        let dictionnary = {
-            let data = (0..dictionnary_len * word_dim)
-                .map(|i| i as i32)
-                .collect::<Vec<_>>();
-            Tensor::construct(data, vec![dictionnary_len, word_dim])
-        };
-        let dict = b.constant(dictionnary);
-        let indexes = b.input(input_shape.to_vec());
-        let res = b.gather(
-            dict,
-            indexes,
-            0,
-            [input_shape.to_vec(), vec![word_dim]].concat(),
-        );
-        b.mark_output(res);
-        b.build()
-    }
-
-    #[test]
-    fn test_gather() {
-        let indices_dims = vec![4, 8];
-        let dict_len = 32;
-        let word_dim = 4;
-        let mut rng = StdRng::seed_from_u64(0x888);
-        let input = Tensor::<i32>::random_range(&mut rng, &indices_dims, 0..dict_len as i32);
-        let model = gather_model(&indices_dims, dict_len, word_dim);
-        unit_test_op(model, &[input]);
-    }
-
-    #[test]
-    fn test_gather_small_path() {
-        let indices_dims = vec![4, 8];
-        let dict_len = 32;
-        let word_dim = 4;
-        let mut rng = StdRng::seed_from_u64(0x889);
-        let input = Tensor::<i32>::random_range(&mut rng, &indices_dims, 0..dict_len as i32);
-        let model = gather_model(&indices_dims, dict_len, word_dim);
-        assert!(matches!(
-            model.graph.nodes[&2].operator,
-            Operator::GatherSmall(_)
-        ));
-        unit_test_op(model, &[input]);
-    }
-
-    #[test]
-    fn test_gather_large_path() {
-        let indices_dims = vec![2, 4];
-        let dict_len = 70000;
-        let word_dim = 2;
-        let mut rng = StdRng::seed_from_u64(0x88A);
-        let input = Tensor::<i32>::random_range(&mut rng, &indices_dims, 0..dict_len as i32);
-        let model = gather_model(&indices_dims, dict_len, word_dim);
-        assert!(matches!(
-            model.graph.nodes[&2].operator,
-            Operator::GatherLarge(_)
-        ));
-        unit_test_op(model, &[input]);
-    }
-
-    #[test]
-    #[ignore = "TODO: non-power-of-two gather path not fully validated yet"]
-    fn test_gather_non_power_of_two_input_len() {
-        let indices_dims = vec![5, 7];
-        let dict_len = 33;
-        let word_dim = 5;
-        let mut rng = StdRng::seed_from_u64(0x889);
-        let input = Tensor::<i32>::random_range(&mut rng, &indices_dims, 0..dict_len as i32);
-        let model = gather_model(&indices_dims, dict_len, word_dim);
-        unit_test_op(model, &[input]);
-    }
-}
